+++ src/main/webapp/static/user.js
@@ -4,13 +4,13 @@ function getUserUrl(){
 	return baseUrl + "/api/supervisor/user";
 }
 
-var updatedId;
 
 //BUTTON ACTIONS
 function addUser(event){
 	//Set the values to update
 	var $form = $("#user-create-form");
 	var json = toJson($form);
+	console.log(json);
 	var url = getUserUrl();
 
 	$.ajax({
@@ -21,7 +21,7 @@ function addUser(event){
        	'Content-Type': 'application/json'
        },	   
 	   success: function(response) {
-	        $('#user-create-modal').modal('toggle');
+	        $('#create-user-modal').modal('toggle');
 	   		getUserList();    
 	   },
 	   error: handleAjaxError
@@ -45,39 +45,27 @@ function getUserList(){
 }
 
 function deleteUser(id){
-	var url = getUserUrl() + "/" + id;
+    var confirmation = confirm("Are You Sure You Want to delete?");
 
-	$.ajax({
-	   url: url,
-	   type: 'DELETE',
-	   success: function(data) {
-	   		getUserList();    
-	   },
-	   error: handleAjaxError
-	});
-}
+    if(confirmation)
+    {
+        var url = getUserUrl() + "/" + id;
 
-function displayEditUser(id){
+        	$.ajax({
+        	   url: url,
+        	   type: 'DELETE',
+        	   success: function(data) {
+        	   		getUserList();
+        	   },
+        	   error: handleAjaxError
+        	});
 
-    updatedId = id;
-
-    var url = getUserUrl() + "/" + updatedId;
-
-    $.ajax({
-    	   url: url,
-    	   type: 'GET',
-    	   success: function(data) {
-    	   		 $("#edit-user-form input[name=email]").val(data.email);
-                 $("edit-user-form input[name=role]").val(data.role);
-
-                     $("#edit-user-modal").modal('toggle');
-    	   },
-    	   error: handleAjaxError
-    	});
-
+    }
+    else{
+        return;
+    }
 
 }
-
 
 
 //UI DISPLAY METHODS
@@ -85,19 +73,20 @@ function displayEditUser(id){
 function displayUserList(data){
 	var $tbody = $('#user-table').find('tbody');
 	$tbody.empty();
+	var sno = 1;
 	for(var i in data){
 		var e = data[i];
 
-		var buttonHtml = '<button style="border:0;padding:0.5rem;background-color:transparent;border:0;border-radius:0.3rem;" title="Edit" onclick="displayEditUser(' + e.id + ')"><i class="fa fa-edit fa-lg"></i></button>&nbsp;&nbsp;&nbsp; '
-		buttonHtml+='<button style="border:0;background-color:transparent;border:0;padding:0.5rem;border-radius:0.3rem;" title="Delete" onclick="deleteUser(' + e.id + ')"><i class="fa fa-trash fa-lg"></i></button>';
+		var buttonHtml = '<button style="border:0;background-color:transparent;border:0;padding:0.5rem;border-radius:0.3rem;" title="Delete" onclick="deleteUser(' + e.id + ')"><i class="fa fa-trash fa-lg"></i></button>';
 
         var row = '<tr>'
-		+ '<td>' + e.id + '</td>'
+		+ '<td>' + sno + '</td>'
 		+ '<td>' + e.email + '</td>'
 		+ '<td>' + e.role + '</td>'
 		+ '<td class="text-center">' + buttonHtml + '</td>'
 		+ '</tr>';
         $tbody.append(row);
+        sno+=1;
 	}
 }
 
@@ -108,35 +97,13 @@ function createUser(){
     $('#create-user-modal').modal('toggle');
 }
 
-function handleEditUser(){
-    var url = getUserUrl() + "/" + updatedId;
 
-    var $form = $("#edit-user-form");
-    var json = toJson($form);
-    console.log(json);
-
-    $.ajax({
-        	   url: url,
-        	   type: 'POST',
-        	   data:json,
-        	   headers: {
-                      	'Content-Type': 'application/json'
-                      },
-        	   success: function() {
-                         $("#edit-user-modal").modal('toggle');
-                         getUserList();
-        	   },
-        	   error: handleAjaxError
-        	});
-
-}
 
 //INITIALIZATION CODE
 function init(){
 	$('#add-user').click(addUser);
 	$('#refresh-data').click(getUserList);
 	$('#create-user').click(createUser);
-	$('#edit-user').click(handleEditUser);
 
 
 }
